requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
click==6.7
colorama==0.3.9
configparser==3.5.0
enum34==1.1.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
PyMySQL==0.7.11
python-dotenv==0.6.5
singledispatch==3.4.0.3
six==1.10.0
urllib==1.21.1
wrapt==1.10.11
